+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Logtail`
   - sends json pretty formatted
   - setUrl() instead of setUri() (old method is deprecated)
+  - `dt` field containing datetime is always properly formatted
 
 ## [1.2.5](https://github.com/orisai/nette-monolog/compare/1.2.4...1.2.5) - 2022-12-14